Added theme selection in settings screen
