+++ src/components/settings/index.ts
@@ -1,0 +1,2 @@
+export * from './theme-section';
+export * from './modules-section';
